move `surround_context/3` to `CodeFragmentUtils`
