+++ apps/language_server/lib/language_server/providers/definition/locator.ex
@@ -25,10 +25,10 @@ defmodule ElixirLS.LanguageServer.Providers.Definition.Locator do
   alias ElixirSense.Core.Parser
 
   alias ElixirLS.LanguageServer.Plugins.Phoenix.Scope
-  alias ElixirLS.LanguageServer.AstUtils
+  alias ElixirLS.LanguageServer.CodeFragmentUtils
 
   def definition(code, line, column, options \\ []) do
-    case AstUtils.surround_context_with_fallback(code, {line, column}) do
+    case CodeFragmentUtils.surround_context_with_fallback(code, {line, column}) do
       {:none, _} ->
         nil
